requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
dateparser==1.2.2
idna==3.11
numpy==2.3.5
pandas==2.3.3
prometheus-api-client==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
regex==2025.11.3
requests==2.32.5
six==1.17.0
tabulate==0.9.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
